Add parse_raw() to handle large XML strings
With large XML strings on Windows the parse function fails to provide the
required functionality due to a path size limitation. Instead of modifying
the parse function to accept a 'type' parameter (which could have values
for 'figure it out', stream, string, filepath, or url) a parse_raw
function is being provided to handle the special case.